+++ RELEASE_NOTES.md
@@ -45,6 +45,7 @@ been tested on Linux. As of this release, the latest known version to work with 
 - (sixlowpan) #1342 Fixed a deserialization error in the MESH header.
 - (dsr) !2762 Fixes header format to comply with RFC4728. Also other minor bug fixes and modernization.
 - (zigbee) In the NWK, broadcast initiator devices are now registered in the BTT to avoid receiving retransmissions.
+- (wifi) !2945 CTS-to-self frames are now transmitted over the bandwidth of the frame they protect, so that transmissions in the rest of the TXOP are no longer limited to 20 MHz.
 
 ## Release 3.48